Day 8 - Advent of Code 2017
This commit contains implementation of maximum registry value calculation
algorythm. This implementation uses Python build in method exec(),
which allows to dynamicly execute Python code provided as a string.

This implementation took more time than usual task on this challenge,
as it was first time ever that I used exec() in a project.